mengubah tipe data object

+++ database/migrations/2024_05_18_130559_create_passengers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('ticket_id');
             $table->enum('category', ['dewasa', 'anak']);
-            $table->decimal('price', 15, 2);
+            $table->integer('price');
             $table->foreign('ticket_id')->references('id')->on('tickets')->onDelete('cascade');
             $table->timestamps();
         });
